requirements.txt
newspaper3k==0.2.8
pandas==1.3.5
scikit-learn==1.0.2
uvicorn==0.19.0
fastapi==0.87.0
gunicorn==20.1.0
joblib==1.2.0
pydantic==1.10.2
gunicorn==20.1.0
pymongo== 4.3.3
python-jose==3.3.0
passlib==1.7.4

